requirements.txt
torch
torchvision
torchaudio
torch-geometric
scikit-learn
numpy
matplotlib
opacus
PyYAML
